Initial Commit
Initial Commit. Rather than importing a JAR of the Qu Client i have been
forced to simply add the source files to the Android project as there
have been issues migrating some of the code to Android (such as
incompatibility with Java 8 and restriction of not being able to do
networking on the main thread). these changes will eventually have to be
made to the stand-alone QuClient project once i'm happy enough that the
QuClient is playing nice with Android.